+++ lib/colored_enumeration/extensions/custom_field_patch.rb
@@ -18,16 +18,24 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA
 
-require 'colored_enumeration'
+module ColoredEnumeration
+  module Extensions
+    module CustomFieldPatch
+      def self.included(base)
+        base.include(InstanceMethods)
+      end
 
-Redmine::Plugin.register :redmine_colored_enumeration do
-  name 'Colored Enumeration'
-  author 'Liane Hampe'
-  description 'Background color for custom field enumeration values'
-  version '0.1.1'
-  url 'https://circle.xmera.de/projects/redmine-colored-enumeration'
-  author_url 'http://xmera.de'
+      module InstanceMethods
+        def color_map
+          enumerations.pluck(:id, :color).to_h
+        end
+      end
+    end
+  end
+end
 
-  requires_redmine version_or_higher: '4.2.1'
-  requires_redmine_plugin :redmine_base_deface, version_or_higher: '1.6.2'
+Rails.configuration.to_prepare do
+  patch = ColoredEnumeration::Extensions::CustomFieldPatch
+  klass = CustomField
+  klass.include patch unless klass.included_modules.include?(patch)
 end